fixed item achievement bugs
