+++ data/items.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-multiple-empty-lines */
 /* eslint-disable space-before-function-paren */
 module.exports = {
-	// CUSTOM EMOJIS  
+	// CUSTOM EMOJIS  🤔
 	'pentagram': {
 		name: 'Pentagram',
 		value: 66666,
